Add data and dom files, change the functions to arrow functions
